chore: ContestParticlesUtils
